# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pharmacy POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/overview:
    get:
      operationId: getDashboardOverview
      tags: [dashboard]
      summary: Get dashboard overview
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Dashboard overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardOverview"
  /dashboard/alerts:
    get:
      operationId: getDashboardAlerts
      tags: [dashboard]
      summary: Get dashboard alerts
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardAlerts"
  /dashboard/top-medicines:
    get:
      operationId: getTopMedicines
      tags: [dashboard]
      summary: Get top-selling medicines
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Top medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopMedicine"
  /dashboard/revenue-trend:
    get:
      operationId: getRevenueTrend
      tags: [dashboard]
      summary: Get revenue trend (daily for last N days)
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Revenue trend
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueTrendPoint"
components:
  schemas:
...
    DashboardOverview:
      type: object
      properties:
        period_revenue:
          type: number
        period_transactions:
          type: integer
        period_patients:
          type: integer
        period_prescriptions:
          type: integer
        avg_daily_revenue:
          type: number
        revenue_growth_pct:
          type: ["number", "null"]
        rx_vs_otc_ratio:
          type: object
          properties:
            rx:
              type: integer
            otc:
              type: integer
          required:
            - rx
            - otc
        retail_vs_hospital:
          type: object
          properties:
            retail:
              type: number
            hospital:
              type: number
          required:
            - retail
            - hospital
        payment_breakdown:
          type: object
          properties:
            cash:
              type: number
            card:
              type: number
            credit:
              type: number
            insurance:
              type: number
          required:
            - cash
            - card
            - credit
            - insurance
      required:
        - period_revenue
        - period_transactions
        - period_patients
        - period_prescriptions
        - avg_daily_revenue
        - rx_vs_otc_ratio
        - retail_vs_hospital
        - payment_breakdown
    DashboardAlerts:
      type: object
      properties:
        low_stock_count:
          type: integer
        expiring_soon_count:
          type: integer
        out_of_stock_count:
          type: integer
        pending_prescriptions:
          type: integer
        overdue_payables:
          type: integer
        expiring_in_7_days:
          type: integer
        expiring_in_30_days:
          type: integer
      required:
        - low_stock_count
        - expiring_soon_count
        - out_of_stock_count
        - pending_prescriptions
        - overdue_payables
        - expiring_in_7_days
        - expiring_in_30_days
    TopMedicine:
      type: object
      properties:
        medicine_id:
          type: string
        medicine_name:
          type: string
        category:
          type: ["string", "null"]
        units_sold:
          type: integer
        revenue:
          type: number
        rank:
          type: integer
      required:
        - medicine_id
        - medicine_name
        - units_sold
        - revenue
        - rank
    RevenueTrendPoint:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        transactions:
          type: integer
        refunds:
          type: number
      required:
        - date
        - revenue
        - transactions
        - refunds